# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type CoachMessageRole = typeof CoachMessageRole[keyof typeof CoachMessageRole];


export const CoachMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface CoachMessage {
  role: CoachMessageRole;
  /**
     * @minLength 1
     * @maxLength 20000
     */
  content: string;
}

export interface CoachHistoryInput {
  /** @maxItems 80 */
  messages: CoachMessage[];
  /** @maxLength 20000 */
  draft: string;
  /** @minimum 0 */
  revision: number;
  /** Original device timestamp, supplied only for explicitly approved import */
  importedAt?: number;
}

export interface CoachHistorySnapshot {
  messages: CoachMessage[];
  draft: string;
  revision: number;
  /** @nullable */
  updatedAt: number | null;
}

export type PlanningKind = typeof PlanningKind[keyof typeof PlanningKind];


export const PlanningKind = {
  character: 'character',
  relationship: 'relationship',
  timeline: 'timeline',
  worldbuilding: 'worldbuilding',
  note: 'note',
} as const;

export interface PlanningFields {
  kind?: PlanningKind;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title?: string;
  /** @maxLength 20000 */
  content?: string;
  /** @maxLength 200 */
  role?: string;
  /** @maxLength 20000 */
  motivations?: string;
  /** @maxLength 20000 */
  traits?: string;
  /** @maxLength 20000 */
  arc?: string;
  /** @maxLength 100 */
  category?: string;
  /** @maxLength 200 */
  timeLabel?: string;
  /**
     * @minimum 1
     * @nullable
     */
  chapterId?: number | null;
  /**
     * @minimum 1
     * @nullable
     */
  characterId?: number | null;
  /**
     * @minimum 1
     * @nullable
     */
  relatedCharacterId?: number | null;
}

export type PlanningInput = PlanningFields & Required<Pick<PlanningFields, 'kind' | 'title'>>;

export type PlanningUpdate = PlanningFields;

export interface PlanningOrderInput {
  kind: PlanningKind;
  ids: number[];
}

export type PlanningRecord = PlanningFields & {
  id: number;
  documentId: number;
  position: number;
  createdAt: string;
  updatedAt: string;
};

export interface HealthStatus {
  status: string;
}

/**
 * Saved planning record
 */
export type PlanningRecordResponse = PlanningRecord;

/**
 * Complete story plan ordered by position then ID
 */
export type PlanningListResponse = PlanningRecord[];