# lib/api-zod/src/generated/types/coachHistoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachMessage } from './coachMessage';

export interface CoachHistoryInput {
  /** @maxItems 80 */
  messages: CoachMessage[];
  /** @maxLength 20000 */
  draft: string;
  /** @minimum 0 */
  revision: number;
  /** Original device timestamp, supplied only for explicitly approved import */
  importedAt?: number;
}