# lib/api-zod/src/generated/types/coachHistorySnapshot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachMessage } from './coachMessage';

export interface CoachHistorySnapshot {
  messages: CoachMessage[];
  draft: string;
  revision: number;
  /** @nullable */
  updatedAt: number | null;
}