# lib/api-zod/src/generated/types/coachMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachMessageRole } from './coachMessageRole';

export interface CoachMessage {
  role: CoachMessageRole;
  /**
     * @minLength 1
     * @maxLength 20000
     */
  content: string;
}